Remove dependencies in The Savvier Seller
