+++ Ani/The Savvier Seller/RegisterInfo.cs
