+++ Ani/The Savvier Seller/The Savvier Seller.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProductVersion>9.0.30729</ProductVersion>
+    <ProductVersion>8.0.30703</ProductVersion>
     <SchemaVersion>2.0</SchemaVersion>
     <ProjectGuid>{D29C455C-53EF-401B-B478-649D8FE84B16}</ProjectGuid>
     <OutputType>Library</OutputType>
@@ -61,6 +61,8 @@
     <DebugType>full</DebugType>
     <PlatformTarget>x86</PlatformTarget>
     <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+    <Optimize>false</Optimize>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|x86'">
     <OutputPath>bin\x86\Release\</OutputPath>
@@ -70,6 +72,7 @@
     <DebugType>pdbonly</DebugType>
     <PlatformTarget>x86</PlatformTarget>
     <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="ani_GalleryShopBase.cs" />
@@ -82,50 +85,57 @@
     <Compile Include="Shopping\ShoppingMeter.cs" />
     <Compile Include="Shopping\ShoppingProgress.cs" />
     <Compile Include="StoreHelperClass.cs" />
+    <Compile Include="ani_GalleryShopRegister.cs" />
+    <Compile Include="CMStoreSet.cs" />
+    <Compile Include="RegisterInfo.cs" />
+    <Compile Include="TaxCollectorSimpleDialog.cs" />
+    <Compile Include="RestockInfo.cs" />
+    <Compile Include="RestockItem.cs" />
+    <Compile Include="RestockItemHelperClass.cs" />
   </ItemGroup>
   <ItemGroup>
-    <Reference Include="0x03D6C8D903CE868C_Sims3GameplaySystems">
-      <HintPath>..\..\Sims3\Compiler\0x03D6C8D903CE868C_Sims3GameplaySystems.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="0x28EE9D383A73463E_mscorlib">
-      <HintPath>..\..\Sims3\Compiler\0x28EE9D383A73463E_mscorlib.dll</HintPath>
-    </Reference>
-    <Reference Include="0x342EE04373CF1E1C_System">
-      <HintPath>..\..\Sims3\Compiler\0x342EE04373CF1E1C_System.dll</HintPath>
-    </Reference>
-    <Reference Include="0x600F9EA1DDC99FB1_Sims3MetaData">
-      <HintPath>..\..\Sims3\Compiler\0x600F9EA1DDC99FB1_Sims3MetaData.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="0x6AC101133051BEF1_System.Xml">
-      <HintPath>..\..\Sims3\Compiler\0x6AC101133051BEF1_System.Xml.dll</HintPath>
-    </Reference>
-    <Reference Include="0x78CF6CF5304D0C4F_ScriptCore">
-      <HintPath>..\..\Sims3\Compiler\0x78CF6CF5304D0C4F_ScriptCore.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="0xB9C90FDC6793BC0A_Sims3GameplayObjects">
-      <HintPath>..\..\Sims3\Compiler\0xB9C90FDC6793BC0A_Sims3GameplayObjects.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="0x0CAE1C361E05B2B3_Sims3StoreObjects">
-      <HintPath>..\..\Sims3\Compiler\0x0CAE1C361E05B2B3_Sims3StoreObjects.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="0xC356DF69B70ADD42_SimIFace">
-      <HintPath>..\..\Sims3\Compiler\0xC356DF69B70ADD42_SimIFace.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="0xF7C3ADE896D4E765_UI">
-      <HintPath>..\..\Sims3\Compiler\0xF7C3ADE896D4E765_UI.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
     <Reference Include="ani_StoreRestockItem">
       <HintPath>ani_StoreRestockItem\bin\x86\Release\ani_StoreRestockItem.dll</HintPath>
     </Reference>
     <Reference Include="ani_StoreSetRegister">
       <HintPath>ani_StoreSetRegister\bin\Release\ani_StoreSetRegister.dll</HintPath>
+    </Reference>
+    <Reference Include="Sims3GameplaySystems">
+      <HintPath>..\..\Sims3\Compiler\0x03D6C8D903CE868C_Sims3GameplaySystems.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="mscorlib">
+      <HintPath>..\..\Sims3\Compiler\0x28EE9D383A73463E_mscorlib.dll</HintPath>
+    </Reference>
+    <Reference Include="System">
+      <HintPath>..\..\Sims3\Compiler\0x342EE04373CF1E1C_System.dll</HintPath>
+    </Reference>
+    <Reference Include="Sims3Metadata">
+      <HintPath>..\..\Sims3\Compiler\0x600F9EA1DDC99FB1_Sims3MetaData.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="System.Xml">
+      <HintPath>..\..\Sims3\Compiler\0x6AC101133051BEF1_System.Xml.dll</HintPath>
+    </Reference>
+    <Reference Include="ScriptCore">
+      <HintPath>..\..\Sims3\Compiler\0x78CF6CF5304D0C4F_ScriptCore.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="Sims3GameplayObjects">
+      <HintPath>..\..\Sims3\Compiler\0xB9C90FDC6793BC0A_Sims3GameplayObjects.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="Sims3StoreObjects">
+      <HintPath>..\..\Sims3\Compiler\0x0CAE1C361E05B2B3_Sims3StoreObjects.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="SimIFace">
+      <HintPath>..\..\Sims3\Compiler\0xC356DF69B70ADD42_SimIFace.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="UI">
+      <HintPath>..\..\Sims3\Compiler\0xF7C3ADE896D4E765_UI.dll</HintPath>
+      <Private>False</Private>
     </Reference>
   </ItemGroup>
   <ItemGroup>
